api/requirements.txt
azure-functions
azure-storage-blob
azure-storage-queue
Pillow
requests
